# pagelanding
Created a simple landing page using HTML and CSS on a theme called Dog shelter. It includes a single navigation bar with a logo attached and also contents like home, services, login, and signup buttons pushed to the right side. Then comes the main division where I created 2 subdivisions called "left" and "right". On the left div, I have put the content regarding the shelter and on the right side the image of a dog with applied CSS properties. Lastly comes the footer and just to apply my knowledge I have added a short message at the center of a div.